feat: Create db only after completed onboarding
  Since we don't need PouchDB until we start synchronizing, we now avoid
  creating the database before the onboarding is completed as it would
  create unnecessary files on disk if the onboarding is never completed.

+++ core/app.js
@@ -90,7 +90,6 @@ class App {
     basePath = path.resolve(basePath)
     this.basePath = findBasePath(basePath)
     this.config = config.load(this.basePath)
-    this.pouch = new Pouch(this.config)
     this.events = new SyncState()
 
     autoBind(this)
@@ -187,7 +186,9 @@ class App {
 
   async removeConfig() {
     log.info('Removing config...')
-    await this.pouch.db.destroy()
+    if (this.pouch) {
+      await this.pouch.db.destroy()
+    }
     for (const name of await fse.readdir(this.basePath)) {
       if (name.startsWith(LOG_BASENAME)) continue
       await fse.remove(path.join(this.basePath, name))
@@ -265,9 +266,15 @@ class App {
     }
 
     const sendPouchDBTree = async () => {
+      if (!this.pouch) {
+        log.warn('no PouchDB so no tree to send')
+        return
+      }
+
       const pouchdbTree = await this.pouch.localTree()
 
       if (!pouchdbTree) {
+        log.warn('no PouchDB tree to send')
         return
       } else {
         return this.uploadFileToSupport(
@@ -302,6 +309,7 @@ class App {
 
   // Instanciate some objects before sync
   instanciate() {
+    this.pouch = new Pouch(this.config)
     this.ignore = Ignore.loadSync(this.config.ignoreRulesPath)
     this.merge = new Merge(this.pouch)
     this.prep = new Prep(this.merge, this.ignore, this.config)
